Time: 0 ms (100%), Space: 42.4 MB (17.15%) - LeetHub

+++ 1922-count-good-numbers/1922-count-good-numbers.java
@@ -2,8 +2,8 @@
 class Solution {
     public int countGoodNumbers(long n) {
         int mod=1000000007;
-        long a=(long)(pow(5,(n+1)/2,mod));
-        long b=(long)(pow(4,n/2,mod));
+        long a=pow(5,(n+1)/2,mod);
+        long b=pow(4,n/2,mod);
         return (int)((a * b) % mod);
     }
     public long pow(long x, long n,int mod) {
